rest of the character dialogue, some sprite changes
- changed and added a few sprites
- added all needed character dialogue (i think)
- added a "new" character (another version of plant alien for when they're near the ship) since i don't think i can move an old one

My computer has been stuttering and freezing lately. I was hoping to commit after finishing the endings, double checking things and apologizing for not saying something about my situation sooner, but I think i should commit this stuff just in case. I've lost a couple things files before and it'd be a shame to lose the work on this.

I may commit the finished ending drawings/gifs in parts due to this issues.